+++ .eslintrc.json
@@ -13,7 +13,14 @@
   "rules": {
     "no-shadow": "off",
     "no-param-reassign": "off",
-    "eol-last": "off"
+    "eol-last": "off",
+    "import/extensions": [
+      1,
+      {
+        "js": "always",
+        "json": "always"
+      }
+    ]
   },
   "ignorePatterns": ["dist/", "build/"]
 }
